backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hsbc-pdf-audit-backend"
version = "0.1.0"
description = "FastAPI backend scaffold for HSBC PDF accessibility audits"
requires-python = ">=3.9"
dependencies = [
  "fastapi==0.115.12",
  "uvicorn==0.30.6",
  "pydantic-settings==2.4.0",
  "openpyxl==3.1.5",
  "pypdf==5.4.0",
  "python-dotenv==1.0.1",
  "requests==2.32.3",
  "httpx==0.27.2"
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.5",
  "pytest-asyncio==0.24.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
